Crate google_vision1

This documentation was generated from Vision crate version 1.0.9+20190628, where 20190628 is the exact revision of the vision:v1 schema built by the mako code generator v1.0.9.

Everything else about the Vision v1 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

files

annotate and async batch annotate

images

annotate and async batch annotate

locations

operations get

operations

cancel, delete, get and list

projects

locations operations get, locations product sets add product, locations product sets create, locations product sets delete, locations product sets get, locations product sets import, locations product sets list, locations product sets patch, locations product sets products list, locations product sets remove product, locations products create, locations products delete, locations products get, locations products list, locations products patch, locations products reference images create, locations products reference images delete, locations products reference images get, locations products reference images list and operations get

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.projects().operations_get(...).doit()
let r = hub.operations().list(...).doit()
let r = hub.locations().operations_get(...).doit()
let r = hub.images().async_batch_annotate(...).doit()
let r = hub.operations().cancel(...).doit()
let r = hub.operations().delete(...).doit()
let r = hub.files().async_batch_annotate(...).doit()
let r = hub.operations().get(...).doit()
let r = hub.projects().locations_operations_get(...).doit()
let r = hub.projects().locations_product_sets_import(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-vision1 = "*"
# This project intentionally uses an old version of Hyper. See
# https://github.com/Byron/google-apis-rs/issues/173 for more
# information.
hyper = "^0.10"
hyper-rustls = "^0.6"
serde = "^1.0"
serde_json = "^1.0"
yup-oauth2 = "^1.0"

A complete example

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_vision1 as vision1;
use vision1::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use vision1::Vision;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Vision::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.operations().list("name")
             .page_token("dolores")
             .page_size(-63)
             .filter("accusam")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

AddProductToProductSetRequest
Request message for the AddProductToProductSet method.

AnnotateFileRequest
A request to annotate one single file, e.g. a PDF, TIFF or GIF file.

AnnotateFileResponse
Response to a single file annotation request. A file may contain one or more images, which individually have their own responses.

AnnotateImageRequest
Request for performing Google Cloud Vision API tasks over a user-provided image, with user-requested features, and with context information.

AnnotateImageResponse
Response to an image annotation request.

AsyncAnnotateFileRequest
An offline file annotation request.

AsyncBatchAnnotateFilesRequest
Multiple async file annotation requests are batched into a single service call.

AsyncBatchAnnotateImagesRequest
Request for async image annotation for a list of images.

BatchAnnotateFilesRequest
A list of requests to annotate files using the BatchAnnotateFiles API.

BatchAnnotateFilesResponse
A list of file annotation responses.

BatchAnnotateImagesRequest
Multiple image annotation requests are batched into a single service call.

BatchAnnotateImagesResponse
Response to a batch image annotation request.

Block
Logical element on the page.

BoundingPoly
A bounding polygon for the detected image annotation.

CancelOperationRequest
The request message for Operations.CancelOperation.

Chunk

Color
Represents a color in the RGBA color space. This representation is designed for simplicity of conversion to/from color representations in various languages over compactness; for example, the fields of this representation can be trivially provided to the constructor of "java.awt.Color" in Java; it can also be trivially provided to UIColor's "+colorWithRed:green:blue:alpha" method in iOS; and, with just a little work, it can be easily formatted into a CSS "rgba()" string in JavaScript, as well.

ColorInfo
Color information consists of RGB channels, score, and the fraction of the image that the color occupies in the image.

ContentRange
Implements the Content-Range header, for serialization only

CropHint
Single crop hint that is used to generate a new crop when serving an image.

CropHintsAnnotation
Set of crop hints that are used to generate new crops when serving images.

CropHintsParams
Parameters for crop hints annotation request.

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

DetectedBreak
Detected start or end of a structural component.

DetectedLanguage
Detected language for a structural component.

DominantColorsAnnotation
Set of dominant colors and their corresponding scores.

DummyNetworkStream

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance:

EntityAnnotation
Set of detected entity features.

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

FaceAnnotation
A face annotation object contains the results of face detection.

Feature
The type of Google Cloud Vision API detection to perform, and the maximum number of results to return for that type. Multiple Feature objects can be specified in the features list.

FileAnnotateCall
Service that performs image detection and annotation for a batch of files. Now only "application/pdf", "image/tiff" and "image/gif" are supported.

FileAsyncBatchAnnotateCall
Run asynchronous image detection and annotation for a list of generic files, such as PDF files, which may contain multiple pages and multiple images per page. Progress and results can be retrieved through the google.longrunning.Operations interface. Operation.metadata contains OperationMetadata (metadata). Operation.response contains AsyncBatchAnnotateFilesResponse (results).

FileMethods
A builder providing access to all methods supported on file resources. It is not used directly, but through the Vision hub.

GcsDestination
The Google Cloud Storage location where the output will be written to.

GcsSource
The Google Cloud Storage location where the input will be read from.

GroupedResult
Information about the products similar to a single product in a query image.

Image
Client image to perform Google Cloud Vision API tasks over.

ImageAnnotateCall
Run image detection and annotation for a batch of images.

ImageAnnotationContext
If an image was produced from a file (e.g. a PDF), this message gives information about the source of that image.

ImageAsyncBatchAnnotateCall
Run asynchronous image detection and annotation for a list of images.

ImageContext
Image context and/or feature-specific parameters.

ImageMethods
A builder providing access to all methods supported on image resources. It is not used directly, but through the Vision hub.

ImageProperties
Stores image properties, such as dominant colors.

ImageSource
External image source (Google Cloud Storage or web URL image location).

ImportProductSetsGcsSource
The Google Cloud Storage location for a csv file which preserves a list of ImportProductSetRequests in each line.

ImportProductSetsInputConfig
The input content for the ImportProductSets method.

ImportProductSetsRequest
Request message for the ImportProductSets method.

InputConfig
The desired input location and metadata.

JsonServerError
A utility type which can decode a server response that indicates error

KeyValue
A product label represented as a key-value pair.

Landmark
A face-specific landmark (for example, a face feature).

LatLng
An object representing a latitude/longitude pair. This is expressed as a pair of doubles representing degrees latitude and degrees longitude. Unless specified otherwise, this must conform to the WGS84 standard. Values must be within normalized ranges.

LatLongRect
Rectangle determined by min and max LatLng pairs.

ListOperationsResponse
The response message for Operations.ListOperations.

ListProductSetsResponse
Response message for the ListProductSets method.

ListProductsInProductSetResponse
Response message for the ListProductsInProductSet method.

ListProductsResponse
Response message for the ListProducts method.

ListReferenceImagesResponse
Response message for the ListReferenceImages method.

LocalizedObjectAnnotation
Set of detected objects with bounding boxes.

LocationInfo
Detected entity location information.

LocationMethods
A builder providing access to all methods supported on location resources. It is not used directly, but through the Vision hub.

LocationOperationGetCall
Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

NormalizedVertex
A vertex represents a 2D point in the image. NOTE: the normalized vertex coordinates are relative to the original image and range from 0 to 1.

Operation
This resource represents a long-running operation that is the result of a network API call.

OperationCancelCall
Starts asynchronous cancellation on a long-running operation. The server makes a best effort to cancel the operation, but success is not guaranteed. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED. Clients can use Operations.GetOperation or other methods to check whether the cancellation succeeded or whether the operation completed despite cancellation. On successful cancellation, the operation is not deleted; instead, it becomes an operation with an Operation.error value with a google.rpc.Status.code of 1, corresponding to Code.CANCELLED.

OperationDeleteCall
Deletes a long-running operation. This method indicates that the client is no longer interested in the operation result. It does not cancel the operation. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED.

OperationGetCall
Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

OperationListCall
Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED.

OperationMethods
A builder providing access to all methods supported on operation resources. It is not used directly, but through the Vision hub.

OutputConfig
The desired output location and metadata.

Page
Detected page from OCR.

Paragraph
Structural unit of text representing a number of words in certain order.

Position
A 3D position in the image, used primarily for Face detection landmarks. A valid Position must have both x and y coordinates. The position coordinates are in the same scale as the original image.

Product
A Product contains ReferenceImages.

ProductSearchParams
Parameters for a product search request.

ProductSearchResults
Results for a product search request.

ProductSet
A ProductSet contains Products. A ProductSet can contain a maximum of 1 million reference images. If the limit is exceeded, periodic indexing will fail.

ProjectLocationOperationGetCall
Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

ProjectLocationProductCreateCall
Creates and returns a new product resource.

ProjectLocationProductDeleteCall
Permanently deletes a product and its reference images.

ProjectLocationProductGetCall
Gets information associated with a Product.

ProjectLocationProductListCall
Lists products in an unspecified order.

ProjectLocationProductPatchCall
Makes changes to a Product resource. Only the display_name, description, and labels fields can be updated right now.

ProjectLocationProductReferenceImageCreateCall
Creates and returns a new ReferenceImage resource.

ProjectLocationProductReferenceImageDeleteCall
Permanently deletes a reference image.

ProjectLocationProductReferenceImageGetCall
Gets information associated with a ReferenceImage.

ProjectLocationProductReferenceImageListCall
Lists reference images.

ProjectLocationProductSetAddProductCall
Adds a Product to the specified ProductSet. If the Product is already present, no change is made.

ProjectLocationProductSetCreateCall
Creates and returns a new ProductSet resource.

ProjectLocationProductSetDeleteCall
Permanently deletes a ProductSet. Products and ReferenceImages in the ProductSet are not deleted.

ProjectLocationProductSetGetCall
Gets information associated with a ProductSet.

ProjectLocationProductSetImportCall
Asynchronous API that imports a list of reference images to specified product sets based on a list of image information.

ProjectLocationProductSetListCall
Lists ProductSets in an unspecified order.

ProjectLocationProductSetPatchCall
Makes changes to a ProductSet resource. Only display_name can be updated currently.

ProjectLocationProductSetProductListCall
Lists the Products in a ProductSet, in an unspecified order. If the ProductSet does not exist, the products field of the response will be empty.

ProjectLocationProductSetRemoveProductCall
Removes a Product from the specified ProductSet.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the Vision hub.

ProjectOperationGetCall
Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

Property
A Property consists of a user-supplied name/value pair.

RangeResponseHeader

ReferenceImage
A ReferenceImage represents a product image and its associated metadata, such as bounding boxes.

RemoveProductFromProductSetRequest
Request message for the RemoveProductFromProductSet method.

ResultType
Information about a product.

ResumableUploadHelper
A utility type to perform a resumable upload from start to end.

SafeSearchAnnotation
Set of features pertaining to the image, computed by computer vision methods over safe-search verticals (for example, adult, spoof, medical, violence).

ServerError

ServerMessage

Status
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. Each Status message contains three pieces of data: error code, error message, and error details.

Symbol
A single symbol representation.

TextAnnotation
TextAnnotation contains a structured representation of OCR extracted text. The hierarchy of an OCR extracted text structure is like this: TextAnnotation -> Page -> Block -> Paragraph -> Word -> Symbol Each structural component, starting from Page, may further have their own properties. Properties describe detected languages, breaks etc.. Please refer to the TextAnnotation.TextProperty message definition below for more detail.

TextProperty
Additional information detected on the structural component.

Vertex
A vertex represents a 2D point in the image. NOTE: the vertex coordinates are in the same scale as the original image.

Vision
Central instance to access all Vision related resource activities

WebDetection
Relevant information for the image from the Internet.

WebDetectionParams
Parameters for web detection request.

WebEntity
Entity deduced from similar images on the Internet.

WebImage
Metadata for online images.

WebLabel
Label to provide extra metadata for the web detection.

WebPage
Metadata for web pages.

Word
A word representation.

XUploadContentType
The X-Upload-Content-Type header.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

UnusedType
Identifies types which are not actually used by the API This might be a bug within the google API schema.

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.