Struct google_vision1::ProjectMethods

pub struct ProjectMethods<'a, C, A> where
    C: 'a,
    A: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on project resources. It is not used directly, but through the Vision hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_vision1 as vision1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use vision1::Vision;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Vision::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `locations_operations_get(...)`, `locations_product_sets_add_product(...)`, `locations_product_sets_create(...)`, `locations_product_sets_delete(...)`, `locations_product_sets_get(...)`, `locations_product_sets_import(...)`, `locations_product_sets_list(...)`, `locations_product_sets_patch(...)`, `locations_product_sets_products_list(...)`, `locations_product_sets_remove_product(...)`, `locations_products_create(...)`, `locations_products_delete(...)`, `locations_products_get(...)`, `locations_products_list(...)`, `locations_products_patch(...)`, `locations_products_reference_images_create(...)`, `locations_products_reference_images_delete(...)`, `locations_products_reference_images_get(...)`, `locations_products_reference_images_list(...)` and `operations_get(...)`
// to build up your call.
let rb = hub.projects();

Methods

impl<'a, C, A> ProjectMethods<'a, C, A>

pub fn locations_products_get(
&self,
name: &str
) -> ProjectLocationProductGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets information associated with a Product.

Possible errors:

Returns NOT_FOUND if the Product does not exist.

Arguments

name - Resource name of the Product to get. Format is: projects/PROJECT_ID/locations/LOC_ID/products/PRODUCT_ID

pub fn locations_products_delete(
&self,
name: &str
) -> ProjectLocationProductDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Permanently deletes a product and its reference images.

Metadata of the product and all its images will be deleted right away, but search queries against ProductSets containing the product may still work until all related caches are refreshed.

Arguments

name - Resource name of product to delete. Format is: projects/PROJECT_ID/locations/LOC_ID/products/PRODUCT_ID

pub fn locations_product_sets_products_list(
&self,
name: &str
) -> ProjectLocationProductSetProductListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the Products in a ProductSet, in an unspecified order. If the ProductSet does not exist, the products field of the response will be empty.

Possible errors:

Returns INVALID_ARGUMENT if page_size is greater than 100 or less than 1.

Arguments

name - The ProductSet resource for which to retrieve Products. Format is: projects/PROJECT_ID/locations/LOC_ID/productSets/PRODUCT_SET_ID

pub fn locations_operations_get(
&self,
name: &str
) -> ProjectLocationOperationGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

Arguments

name - The name of the operation resource.

pub fn locations_product_sets_delete(
&self,
name: &str
) -> ProjectLocationProductSetDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Permanently deletes a ProductSet. Products and ReferenceImages in the ProductSet are not deleted.

The actual image files are not deleted from Google Cloud Storage.

Arguments

name - Resource name of the ProductSet to delete. Format is: projects/PROJECT_ID/locations/LOC_ID/productSets/PRODUCT_SET_ID

pub fn locations_product_sets_remove_product(
&self,
request: RemoveProductFromProductSetRequest,
name: &str
) -> ProjectLocationProductSetRemoveProductCall<'a, C, A>

Create a builder to help you perform the following task:

Removes a Product from the specified ProductSet.

Arguments

request - No description provided.

name - The resource name for the ProductSet to modify. Format is: projects/PROJECT_ID/locations/LOC_ID/productSets/PRODUCT_SET_ID

pub fn locations_products_list(
&self,
parent: &str
) -> ProjectLocationProductListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists products in an unspecified order.

Possible errors:

Returns INVALID_ARGUMENT if page_size is greater than 100 or less than 1.

Arguments

parent - The project OR ProductSet from which Products should be listed. Format: projects/PROJECT_ID/locations/LOC_ID

pub fn operations_get(&self, name: &str) -> ProjectOperationGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

Arguments

name - The name of the operation resource.

pub fn locations_products_create(
&self,
request: Product,
parent: &str
) -> ProjectLocationProductCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates and returns a new product resource.

Possible errors:

Returns INVALID_ARGUMENT if display_name is missing or longer than 4096 characters.

Returns INVALID_ARGUMENT if description is longer than 4096 characters.

Returns INVALID_ARGUMENT if product_category is missing or invalid.

Arguments

request - No description provided.

parent - The project in which the Product should be created. Format is projects/PROJECT_ID/locations/LOC_ID.

pub fn locations_product_sets_get(
&self,
name: &str
) -> ProjectLocationProductSetGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets information associated with a ProductSet.

Possible errors:

Returns NOT_FOUND if the ProductSet does not exist.

Arguments

name - Resource name of the ProductSet to get. Format is: projects/PROJECT_ID/locations/LOG_ID/productSets/PRODUCT_SET_ID

pub fn locations_product_sets_create(
&self,
request: ProductSet,
parent: &str
) -> ProjectLocationProductSetCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates and returns a new ProductSet resource.

Possible errors:

Returns INVALID_ARGUMENT if display_name is missing, or is longer than 4096 characters.

Arguments

request - No description provided.

parent - The project in which the ProductSet should be created. Format is projects/PROJECT_ID/locations/LOC_ID.

pub fn locations_product_sets_add_product(
&self,
request: AddProductToProductSetRequest,
name: &str
) -> ProjectLocationProductSetAddProductCall<'a, C, A>

Create a builder to help you perform the following task:

Adds a Product to the specified ProductSet. If the Product is already present, no change is made.

One Product can be added to at most 100 ProductSets.

Possible errors:

Returns NOT_FOUND if the Product or the ProductSet doesn't exist.

Arguments

request - No description provided.

name - The resource name for the ProductSet to modify. Format is: projects/PROJECT_ID/locations/LOC_ID/productSets/PRODUCT_SET_ID

pub fn locations_product_sets_import(
&self,
request: ImportProductSetsRequest,
parent: &str
) -> ProjectLocationProductSetImportCall<'a, C, A>

Create a builder to help you perform the following task:

Asynchronous API that imports a list of reference images to specified product sets based on a list of image information.

The google.longrunning.Operation API can be used to keep track of the progress and results of the request. Operation.metadata contains BatchOperationMetadata. (progress) Operation.response contains ImportProductSetsResponse. (results)

The input source of this method is a csv file on Google Cloud Storage. For the format of the csv file please see ImportProductSetsGcsSource.csv_file_uri.

Arguments

request - No description provided.

parent - The project in which the ProductSets should be imported. Format is projects/PROJECT_ID/locations/LOC_ID.

pub fn locations_products_patch(
&self,
request: Product,
name: &str
) -> ProjectLocationProductPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Makes changes to a Product resource. Only the display_name, description, and labels fields can be updated right now.

If labels are updated, the change will not be reflected in queries until the next index time.

Possible errors:

Returns NOT_FOUND if the Product does not exist.

Returns INVALID_ARGUMENT if display_name is present in update_mask but is missing from the request or longer than 4096 characters.

Returns INVALID_ARGUMENT if description is present in update_mask but is longer than 4096 characters.

Returns INVALID_ARGUMENT if product_category is present in update_mask.

Arguments

request - No description provided.

name - The resource name of the product. Format is: projects/PROJECT_ID/locations/LOC_ID/products/PRODUCT_ID. This field is ignored when creating a product.

pub fn locations_product_sets_patch(
&self,
request: ProductSet,
name: &str
) -> ProjectLocationProductSetPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Makes changes to a ProductSet resource. Only display_name can be updated currently.

Possible errors:

Returns NOT_FOUND if the ProductSet does not exist.

Returns INVALID_ARGUMENT if display_name is present in update_mask but missing from the request or longer than 4096 characters.

Arguments

request - No description provided.

name - The resource name of the ProductSet. Format is: projects/PROJECT_ID/locations/LOC_ID/productSets/PRODUCT_SET_ID. This field is ignored when creating a ProductSet.

pub fn locations_product_sets_list(
&self,
parent: &str
) -> ProjectLocationProductSetListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists ProductSets in an unspecified order.

Possible errors:

Returns INVALID_ARGUMENT if page_size is greater than 100, or less than 1.

Arguments

parent - The project from which ProductSets should be listed. Format is projects/PROJECT_ID/locations/LOC_ID.

pub fn locations_products_reference_images_create(
&self,
request: ReferenceImage,
parent: &str
) -> ProjectLocationProductReferenceImageCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates and returns a new ReferenceImage resource.

The bounding_poly field is optional. If bounding_poly is not specified, the system will try to detect regions of interest in the image that are compatible with the product_category on the parent product. If it is specified, detection is ALWAYS skipped. The system converts polygons into non-rotated rectangles.

Note that the pipeline will resize the image if the image resolution is too large to process (above 50MP).

Possible errors:

Returns INVALID_ARGUMENT if the image_uri is missing or longer than 4096 characters.

Returns INVALID_ARGUMENT if the product does not exist.

Returns INVALID_ARGUMENT if bounding_poly is not provided, and nothing compatible with the parent product's product_category is detected.

Returns INVALID_ARGUMENT if bounding_poly contains more than 10 polygons.

Arguments

request - No description provided.

parent - Resource name of the product in which to create the reference image. Format is projects/PROJECT_ID/locations/LOC_ID/products/PRODUCT_ID.

pub fn locations_products_reference_images_delete(
&self,
name: &str
) -> ProjectLocationProductReferenceImageDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Permanently deletes a reference image.

The image metadata will be deleted right away, but search queries against ProductSets containing the image may still work until all related caches are refreshed.

The actual image files are not deleted from Google Cloud Storage.

Arguments

name - The resource name of the reference image to delete. Format is: projects/PROJECT_ID/locations/LOC_ID/products/PRODUCT_ID/referenceImages/IMAGE_ID

pub fn locations_products_reference_images_get(
&self,
name: &str
) -> ProjectLocationProductReferenceImageGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets information associated with a ReferenceImage.

Possible errors:

Returns NOT_FOUND if the specified image does not exist.

Arguments

name - The resource name of the ReferenceImage to get. Format is: projects/PROJECT_ID/locations/LOC_ID/products/PRODUCT_ID/referenceImages/IMAGE_ID.

pub fn locations_products_reference_images_list(
&self,
parent: &str
) -> ProjectLocationProductReferenceImageListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists reference images.

Possible errors:

Returns NOT_FOUND if the parent product does not exist.

Returns INVALID_ARGUMENT if the page_size is greater than 100, or less than 1.

Arguments

parent - Resource name of the product containing the reference images. Format is projects/PROJECT_ID/locations/LOC_ID/products/PRODUCT_ID.